Helper-project/Magnus AI/apps/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.32
alembic==1.13.2
pydantic==2.8.2
python-dotenv==1.0.1
httpx==0.27.0
python-chess==1.999
